+++ CHANGELOG.md
@@ -2,3 +2,4 @@
 
 1. Setup Folder Structure and loaded Starter Code
 2. Added scripts to container package.json & added htmlWebPackPlugi to webpack.config.js
+3. Updated package.json, index.js & webpack.config.json